MAT-8779 add buttons for table manipulation
