+++ react/RichTextEditor/tableIcons/tabelIconsIndex.js
@@ -1,0 +1,17 @@
+import DeleteTableIcon from "./deleteTableIcon";
+import AddRowBelowIcon from "./addRowBelowIcon";
+import AddRowAboveIcon from "./addRowAboveIcon";
+import DeleteRowIcon from "./deleteRowIcon";
+import AddColumnLeftIcon from "./addColumnLeftIcon";
+import AddColumnRightIcon from "./addColumnRightIcon";
+import DeleteColumnIcon from "./deleteColumnIcon"; 
+
+export {
+  DeleteTableIcon,
+  AddRowBelowIcon,
+  AddRowAboveIcon,
+  DeleteRowIcon,
+  AddColumnLeftIcon,
+  AddColumnRightIcon,
+  DeleteColumnIcon,
+};
